fully functioning save system+fantanill
I also added the timer.

you have to input a file directly into the path slot for the save data object under main cammera for it to work; but once you do that you will get a json file in that folder that says "island_1_1.json" which will allow you to then open it after saving with a fantanill clone sitting in the project and it will actively show the internal structure on how he is placed

grid building system by tamara games as used here.

I hope to get a element combination system, shop setup, and possibly a background done soon for which then i will just have to then add every creature for neglected nostalgia. to say that i have a working demo. i will not release it (and nor can you due to copyright :)   ) untill i get it so it will both save the file folder so you dont have to put it in everytime; and a map system so you can go between the places and possibly series. (again series are the groups of creatures, and the places are the story, or place that the creatures go through. only some series are stationary.)