shop system and save system fix
i fixed a loading bug where if you spammed load you would get duplicates of everything you currently got, also to work the shop you need to use the "savedata" object (it is in the dropdown menu of the main cammera object) in the editor while in play and give yourself more coins in the "Coin_editable" varible, and click save which will update the coins value to then use in the shop. (this is a debug tool.) you need more then 100 coins hoewevere for the fantanill button to do anything in the shop. (i will add more ui elements; but as of right now everything is temporary and is bound to change.)