fully implemented carreeon
I also made it so they save whether they are asleep or not, that way when you go about an reload the game with some asleep they will remain asleep.
(I do have to reimplement the sleep system though, but i will try and make it better now!) at the same time i'll also be adding virteghoul and a breeding system so you can obtain her without spending diamonds, which while on that i may change "diamonds" to be aesthetically(viewed by player) "gemstones", i will not change scripts to reflect this though as that'd be a pain to do, i may also swap "coins" for "gold" in the same way.