edit to last one
you can now move the creatures during runtime and save, and when it loads the creatures will be in their new positions.
same for sleep and orientation. all that's left in that regard is the interaction system to reassemble them as the temp for the building system, or a sleep button.

i may also redo the currency collection system a bit when i add the interaction system.