worked on msc live #1
just made the quagmire workable, although it does need some fine tuning.

(I will most indefinitely make a scene template for making a new place as what i went through pre-livestream and suring most of it was a nigtmare to flip between and through.)

I do hope to make it so it automatically saves the game like every 10 seconds.