here is the end of the stream i just did
i also made it so if you cancel the moving the creature while moving it it wont continue to follow the cursor as you click about. i will make ui for moving the creatures now! (i will be using the canvas for world space to have it move to the creature when you click it and will dissapear when moving it about. this will also make it known what creature you have clicked and you wont have to do geusswork.)