making my own font + fixed ec
just fixed an elemental combiner bug that was just a forget to reset it when loading island since it would need todo that every time it loaded and not once. also i made my own font for this game, and it turned out amazing, though a bit missmatched, but i like it like that since its supposed to be my handwriting. (i did one letter at a time and scaled it down.)