song skip system
it needs to be filled in manually for every bar of the song, but once done, it will skip over parts where the creature is asleep.

it is broken currently regarding the bass as it is set up in neglected nostalgia to skip if it is in, but due to me not including the base into the overall for the creatures it does break in that regard. (I plan on doing something like the lost landscapes did for evergreen marsh where it actually comes from something, and in this case it also allows me to make a specific creature or thing that will only play the bass. I will work on that thing tommorrow, though i do know for at least the dreamicks it will be a little speaker and its off state will have it not vibrate or make noise!) while on that i will also rework the animation system a bit so that like when the creature goes to sleep, when it goes into idle it will turn off the audio source.