null setup
the "_null_creatureData" objects are so that when put into a place with blanks between the datas (as the elemental combiner is heavy on that being percise) it can work without any extra function. they also have the designation of -1 in the id slot so the shop skips over it when automatically setting up.