mrs incubator added
now when you get a creature, it will be sent to mrs incubator before you can actually place it.

(she is indeed the msc version of the incubator from my singing mosnters. i still dohave to add the "done" ui image, but aside from that she is here!)

i've also completely redone how the storefront works with this, though it is mostly how it deals with the output. now i need to make a descission on whether or not to include extra structures or even decorations.